Added experiment type to all_editedDataIDs stored procedure Dropped two constraints in noise and HPCRequestData tables Added check for Qt 4 environment in makeall.sh Removed references to QCA in *.pri files Fixed us_rotor_main to let the gui delete itself

+++ makeall.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+QMAKEVER=`qmake --version|grep version|grep 4`
+if [ "x$QMAKEVER" == "x" ]; then
+  echo "Wrong qmake, check environment"
+  exit
+fi
+
 DIR=$(pwd)
 rm -f build.log
 NBERR=0
